web/Cross-College-Collage-Compitition/includes/requirements.txt
beautifulsoup4
bs4
lxml==4.6.3
flask
gunicorn
requests
urllib3
pillow
Werkzeug
python-magic
selenium
flask-limiter